Create CarSpeedDetector
The ultimate goal of the project is to prevent accidents in campus premises caused by vehicles exceeding the speed limit. This is an important issue, as campus premises are areas where the safety of students and others should be a top priority. To address this issue, the system captures the speed of passing vehicles and displays it, allowing drivers to be aware of their speed and make adjustments if necessary. The system is equipped with 3 IR sensors and an Arduino-based system, which is used to process the data from the sensors and display the speed on a display unit. If a vehicle is found to be exceeding the speed limit, the system gives a warning message to alert the driver to slow down, helping to ensure the safety of those within the university campus